Preserve people's positions at end of simulation
People's positions are calculated based on the current stopwatch time.
Before, at the end of the sim this time was being reset to 0, which
caused a "jump" in the display as all the dots moved to different
positions. Now, at the end of the sim the time is preserved, so the
display keeps people in the right spots.